Conflicts:
	TaskDoc/src/main/java/com/taskdoc/www/database/dao/chatroomjoin/ChatRoomJoinDAO.java
	TaskDoc/src/main/java/com/taskdoc/www/database/dao/chatroomjoin/ChatRoomJoinDAOImpl.java
	TaskDoc/src/main/webapp/WEB-INF/views/tetete.jsp
	TaskDoc/target/classes/com/taskdoc/www/database/dao/chatroomjoin/ChatRoomJoinDAO.class
	TaskDoc/target/classes/com/taskdoc/www/database/dao/chatroomjoin/ChatRoomJoinDAOImpl.class
	TaskDoc/target/m2e-wtp/web-resources/META-INF/maven/com.taskdoc/www/pom.properties
2018.08.29 merge

+++ TaskDoc/src/main/webapp/WEB-INF/views/user/infoEdit.jsp
@@ -1,50 +1,186 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!DOCTYPE HTML>
 <html>
 <head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 
-<title>Insert title here</title>
-</head>
-<body>
-	<button onclick="testBtn">버턴</button>
-</body>
+
+<%@include file="/WEB-INF/views/fix/header.jsp"%>
+<link rel="stylesheet"
+	href="${pageContext.request.contextPath }/resources/css/bootstrap/css/bootstrap.css">
+<%
+	String loginid = "";
+	loginid = (String) session.getAttribute("loginid");
+	String paramId=request.getParameter("id");
+%>
 
 <script type="text/javascript">
-	/*회원 정보 수정을위해 정보를 불러오기*/
-	/*response는 list형태로 return*/
-	function testBtn() {
+var id='<%=loginid%>';
+	if (id == "null") {
+		alert('로그인이 필요한 페이지입니다.');
+		window.location.href = '/';
+	}
+</script>
+</head>
+<body>
+	<!--FRAME  -->
+	<div id="frame">
+
+		<!--MAIN HEADER  -->
+		<%@include file="/WEB-INF/views/fix/main_header.jsp"%>
+		<!--MAIN HEADER  -->
+
+		<!--WRAPPER  -->
+		<div id="wrapper">
+
+			<!--SIDE BAR  -->
+			<%@include file="/WEB-INF/views/fix/left_side.jsp"%>
+			<!--SIDE BAR  -->
+
+			<!--CONTENTWRAP  -->
+			<div id="contentwrap" style="background-color: #e0e0e0;">
+				<div class="container" style="display: table; vertical-align: middle;">
+					<div class="main"
+						style="border: 1px solid rgba(0, 0, 0, 0.2); box-shadow: 0 3px 9px rgba(0, 0, 0, 0.5); background-color: white; display: table-cell;">
+						<div id="1"
+							style="width: 100%; border-bottom: 1px solid rgba(0, 0, 0, 0.2);">
+							<span id="u" style="margin-left: 20px; font-size: 20px;">회원
+								정보 수정</span>
+						</div>
+						<div id="2" class="bts" style="width: 100%; margin-top: 10px;">
+							<span>아이디</span> <input id="uid"
+								style="width: 80%; margin-top: 5px;" type="text"
+								class="form-control" readonly="readonly">
+						</div>
+						<div id="3" class="bts" style="width: 100%;">
+							<span>비밀번호</span> <input id="upasswd"
+								style="width: 80%; margin-top: 5px;" type="text"
+								class="form-control" readonly="readonly">
+						</div>
+						<div id="4" class="bts" style="width: 100%;">
+							<span>이름</span> <input id="uname"
+								style="width: 80%; margin-top: 5px;" type="text"
+								class="form-control">
+						</div>
+						<div id="5" class="bts" style="width: 100%;">
+							<span>상태메세지</span> <input id="ustate"
+								style="width: 80%; margin-top: 5px;" type="text"
+								class="form-control">
+						</div>
+						<div id="6" class="bts" style="width: 100%;">
+							<span>휴대폰번호(' - ' 는 빼주시고 입력해주세요.)</span> <input id="uphone"
+								style="width: 80%; margin-top: 5px;" type="text"
+								class="form-control">
+						</div>
+						<div id="7"
+							style="width: 100%; border-top: 1px solid rgba(0, 0, 0, 0.2);">
+							<div style="width: 100%; height: 30%;">
+								<span id="userinfor"
+									style="float: right; margin-right: 5px; margin-top: 5px; color: rgba(0, 0, 0, 0.8);'">TASKDOC
+									(은)는 개인정보 방침을 준수합니다.</span>
+							</div>
+
+							<div class="bts" style="height: 70%; margin-top: 16px;">
+								<button type="button"
+									style="margin-left: 7px; float: left; background-color: #ed8151; border: 0; outline: none;"
+									class="btn btn-success" onclick="userdel()">회원탈퇴</button>
+								<button type="button"
+									style="float: right; margin-right: 7px; background-color: #ed8151; border: 0; outline: none;"
+									class="btn btn-success" onclick="cancel()">Cancel</button>
+								<button type="button" class="btn btn-success" onclick="edit()"
+									style="margin: auto 0; margin-left: 6px; float: right; margin-right: 7px; background-color: #ed8151; border: 0; outline: none;">Edit</button>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+			<!--PROJECT_WRAP -->
+
+		</div>
+		<!--CONTENTWRAP  -->
+
+		<!-- FOOTER -->
+		<%@include file="/WEB-INF/views/fix/footer.jsp"%>
+		<!-- FOOTER -->
+	</div>
+	<!--FRAME  -->
+
+</body>
+
+
+<script>
+var savepasswd="";
+
+	$(document).ready(function() {
+		var contentwWidth = $("#contentwrap").width();
+		var contentHeight = $("#contentwrap").height();
+
+		var mainwidth = parseFloat(contentwWidth / 2.4);
+		$(".main").css('width', mainwidth);
+
+		var mainheight = parseFloat(contentHeight / 1.3);
+		$(".main").css('height', mainheight);
+
+		$("#1").css('height', mainheight / 7 + "px");
+		$("#1").css('line-height', mainheight / 7 + "px");
+		for (var i = 2; i < 7; i++) {
+			$("#" + i).css('height', mainheight / 7 + "px");
+			$("#" + i).css('margin-left', "20px");
+		}
+		$("#7").css('height', mainheight / 7 + "px");
+
+		var buttonwidth = parseFloat(mainheight / 3 + "px");
+		$("#a").css('width', buttonwidth + "px");
+		$("#b").css('width', buttonwidth + "px");
+		$("#c").css('width', buttonwidth + "px");
+
+		var margintop = parseFloat(contentHeight / 10);
+		$(".container").css('margin-top', margintop + "px");
+		
+		/*회원 정보 수정을위해 정보를 불러오기*/
 		$.ajax({
 			type : 'GET',
-			url : 'userinfo/' + '회원ID',
+			url : '/userinfo/' + '<%=paramId%>',
 			success : function(response) {
-				if (response.length==0) {
-					alert('Server or Client ERROR, 회원정보 불러오기 실패');
-				} else if(response.length!=0) {
-					alert('Server or Client ERROR, 회원정보 불러오기 성공');
-
+				//Object.keys(response).length --> 자바리턴값이 Object일때 Object 의 크기 구하는법
+				if (Object.keys(response).length > 0) {
+					alert('유저정보 조회완료');
+					savepasswd=response.upasswd;
+					$('#uid').val(response.uid);
+					$('#upasswd').val(response.upasswd);
+					$('#uname').val(response.uname);
+					$('#ustate').val(response.ustate);
+					$('#uphone').val(response.uphone);
+				} else {
+					alert('유저정보 조회 실패')
 				}
 			},
 			error : function(e) {
 				alert("ERROR : " + e.statusText);
 			}
 		});
-		/*/회원 정보 수정을위해 정보를 불러오기*/
-	}
+	});
+	/*/회원 정보 수정을위해 정보를 불러오기*/
 	
-	/* 회원 정보 수정 처리*/
+	
+
+/* 회원 정보 수정 처리*/
+function edit() {
+	var param = {
+		'uid' : '<%=loginid%>',
+		'upasswd':savepasswd,
+		'uname' : $("#uname").val(),
+		'ustate' : $("#ustate").val(),
+		'uphone' : $("#uphone").val()
+	};
 	$.ajax({
 		type : 'PUT',
-		url : 'userinfo/',
+		url : '/userinfo',
 		contentType : 'application/json',
 		data : JSON.stringify(param),
 		success : function(response) {
-			/* response 1 or -1
-			 */
-			if (response == 1) {
-				alert('회원 정보 수정 완료');
+			if (response >0) {
+				alert('회원정보 수정 완료!');
 			} else {
 				alert('Server or Client ERROR, 회원정보 수정 실패');
 			}
@@ -53,26 +189,34 @@
 			alert("ERROR : " + e.statusText);
 		}
 	});
-	/*/회원 정보 수정 처리*/
+}
+/*/회원 정보 수정 처리*/
 
-	/* 회원 탈퇴*/
-	$.ajax({
-		type : 'DELETE',
-		url : 'userinfo/' + '로그인한 ID',
-		success : function(response) {
-			/* response 1 or -1
-			 */
-			if (response == 1) {
-				alert('회원  탈퇴 완료');
-			} else {
-				alert('Server or Client ERROR, 회원 탈퇴 실패');
+function userdel(){
+	if(confirm("※회원 탈퇴를 하시겠습니까?")==true){
+		$.ajax({
+			type : 'DELETE',
+			url : '/userinfo/' + '<%=loginid%>',
+			success : function(response) {
+				/* response 1 or -1
+				 */
+				if (response >0) {
+					alert('회원  탈퇴 완료');
+					window.location.href='/';
+				} else {
+					alert('Server or Client ERROR, 회원 탈퇴 실패');
+				}
+			},
+			error : function(e) {
+				alert("ERROR : " + e.statusText);
 			}
-		},
-		error : function(e) {
-			alert("ERROR : " + e.statusText);
-		}
-	});
-	/*/회원 탈퇴*/
-	
+		});
+	}else{
+		return;
+	}
+}
+function cancel() {
+	location.href = '/project/main';
+}
 </script>
 </html>
